+++ src/microservices/Microservices.CohortExtractor/Execution/CohortExtractorHost.cs
@@ -132,6 +132,9 @@ namespace Microservices.CohortExtractor.Execution
             if(!string.IsNullOrWhiteSpace(_consumerOptions.RejectorType))
                 _fulfiller.Rejectors.Add(ObjectFactory.CreateInstance<IRejector>(_consumerOptions.RejectorType,typeof(IRejector).Assembly));
 
+            if(_consumerOptions.RejectPatientsIn.HasValue)
+                _fulfiller.Rejectors.Add(new PatientRejector(repositoryLocator.CatalogueRepository.GetObjectByID<ColumnInfo>(_consumerOptions.RejectPatientsIn.Value)));
+
             if(_consumerOptions.Blacklists != null)
                 foreach (int id in _consumerOptions.Blacklists)
                 {